Fixed the RK4 integrator, gave new look to graph
The Runge-Kutta 4 integrator wasn't actually RK4, so that has been fixed. Also, changed the code a bit to make more sense in the context (instead of just hardcoding the movement equations into the solver, pass a function into the solver). It also now draws the actual path of the normal equation (since we can just integrate by hand).